+++ src/features/movies/MoviePage/components/Backdrop/styled.js
