+++ src/features/movies/MoviePage/index.js
@@ -1,7 +1,7 @@
 import React from "react";
-import MovieTile from "./MovieTile";
-import Backdrop from "./Backdrop";
-import MovieCast from "./MovieCast";
+import MovieTile from "./components/MovieTile";
+import Backdrop from "./components/Backdrop";
+import MovieCast from "./components/MovieCast";
 
 const MoviePage = () => (
     <>
